feat: parse mustache expressions inside attribute value strings
Not supporting this was reported in #2

Note that we use a short-cut here to enable mustache statements inside
of a string that disallows a `{` to appear inside a "concat statement".

Technically this should be allowed, though I would say it's quite rare.

Issue #6 was files to keep track of eventually fixing that, though it
will likely require moving the scanner into C to pull that off.

The problem is that `{` alone is safe, but `{{` together is not, and I'm
not sure you can write a regex that TreeSitter supports that can capture
that correctly.