Native support for else and else if in block statements
As there isn't a native concept of if statements in the grammar, it
makes it especially hard to query `else if` statements, as they've
previously just been exposed as a helper. It makes it cumbersome to
query `else` blocks and not possible—I think at least—to query and
capture `else if` as the same capture.

`else` is supported for all block_statements in Ember, but `else if`
isn't, but it wouldn't be valid syntax that we want to capture any other
way, so I haven't tried to be cleaver about it.